Fixed the CourseDAO so that the query for courses eligible for enrollment now works properly.
Added a .equals() and .hashCode() method to the User class so it can
support Set operations required for some comparisons on the CoureDAO
class.

Updated the MockDataGenerator to only create its sample Courses if they
do not already exist.  My instance was building up a lot of dummy data
that was the same.

#32